WIP Move configmap watcher to its own package
