+++ pkg/configmap/configmap_watcher_test.go
@@ -1,4 +1,4 @@
-package controller
+package configmap
 
 import (
 	"context"
